Crate ts3plugin

Modules

ts3interface

Macros

create_plugin!
Create a plugin. This macro has to be called once per library to create the function interface that is used by TeamSpeak.

Structs

PluginData

Enums

InitResult

Traits

Plugin
The trait that has to be implemented by a plugin. To enhance a library to a working TeamSpeak plugin you have to call the macro create_plugin! afterwards.