Trait ts3plugin::plugin::Plugin

pub trait Plugin: 'static + Send {
    fn new(api: &mut TsApi) -> Result<Box<Self>, InitError> where Self: Sized;

    fn connect_status_change(&mut self, api: &mut TsApi, server_id: ServerId, status: ConnectStatus, error: Error) { ... }
    fn server_stop(&mut self, api: &mut TsApi, server_id: ServerId, message: String) { ... }
    fn server_error(&mut self, api: &mut TsApi, server_id: ServerId, error: Error, message: String, return_code: String, extra_message: String) -> bool { ... }
    fn server_edited(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker) { ... }
    fn server_connection_info(&mut self, api: &mut TsApi, server_id: ServerId) { ... }
    fn connection_info(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId) { ... }
    fn connection_updated(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_connection: Option<Connection>, invoker: Invoker) { ... }
    fn connection_announced(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, appeared: bool) { ... }
    fn connection_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, connected: bool, message: String) { ... }
    fn connection_move(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility) { ... }
    fn connection_moved(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility, invoker: Invoker) { ... }
    fn connection_timeout(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId) { ... }
    fn channel_announced(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId) { ... }
    fn channel_description_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId) { ... }
    fn channel_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, old_channel: Option<Channel>) { ... }
    fn channel_created(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, invoker: Invoker) { ... }
    fn channel_deleted(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, invoker: Invoker) { ... }
    fn channel_edited(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, old_channel: Option<Channel>, invoker: Invoker) { ... }
    fn channel_password_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId) { ... }
    fn channel_moved(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, new_parent_channel_id: ChannelId, invoker: Invoker) { ... }
    fn message(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker, target: MessageReceiver, message: String, ignored: bool) -> bool { ... }
    fn poke(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker, message: String, ignored: bool) -> bool { ... }
    fn channel_kick(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility, invoker: Invoker, message: String) { ... }
    fn server_kick(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, invoker: Invoker, message: String) { ... }
    fn server_ban(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, invoker: Invoker, message: String, time: u64) { ... }
    fn talking_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, talking: TalkStatus, whispering: bool) { ... }
    fn avatar_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, path: Option<String>) { ... }
    fn connection_channel_group_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, channel_group_id: ChannelGroupId, channel_id: ChannelId, invoker: Invoker) { ... }
    fn connection_server_group_added(&mut self, api: &mut TsApi, server_id: ServerId, connection: Invoker, server_group_id: ServerGroupId, invoker: Invoker) { ... }
    fn connection_server_group_removed(&mut self, api: &mut TsApi, server_id: ServerId, connection: Invoker, server_group_id: ServerGroupId, invoker: Invoker) { ... }
    fn playback_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, samples: &mut [i16], channels: i32) { ... }
    fn post_process_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, samples: &mut [i16], channels: i32, channel_speaker_array: &[Speaker], channel_fill_mask: &mut u32) { ... }
    fn mixed_playback_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, samples: &mut [i16], channels: i32, channel_speaker_array: &[Speaker], channel_fill_mask: &mut u32) { ... }
    fn captured_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, samples: &mut [i16], channels: i32, send: &mut bool) -> bool { ... }
    fn permission_error(&mut self, api: &mut TsApi, server_id: ServerId, permission_id: PermissionId, error: Error, message: String, return_code: String) -> bool { ... }
    fn shutdown(&mut self, api: &mut TsApi) { ... }
}

The trait that has to be implemented by a plugin. To enhance a library to a working TeamSpeak plugin you have to call the macro create_plugin! afterwards.

Required Methods

fn new(api: &mut TsApi) -> Result<Box<Self>, InitError> where Self: Sized

Called when the plugin is loaded by TeamSpeak.

Provided Methods

fn connect_status_change(&mut self, api: &mut TsApi, server_id: ServerId, status: ConnectStatus, error: Error)

If the connection status changes. If status = ConnectStatus::Connecting, the connection_id is not yet registered in the TsApi.

fn server_stop(&mut self, api: &mut TsApi, server_id: ServerId, message: String)

fn server_error(&mut self, api: &mut TsApi, server_id: ServerId, error: Error, message: String, return_code: String, extra_message: String) -> bool

Return false if the TeamSpeak client should handle the error normally or true if the client should ignore the error.

fn server_edited(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker)

fn server_connection_info(&mut self, api: &mut TsApi, server_id: ServerId)

fn connection_info(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId)

fn connection_updated(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_connection: Option<Connection>, invoker: Invoker)

fn connection_announced(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, appeared: bool)

If the plugin was informed about a new connection. If appeared is true, the connection was previously not known to the plugin, if appeared is false, the connection left the view of connection.

fn connection_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, connected: bool, message: String)

Called, if a connection connects to the server. This is also called for our own connection.

fn connection_move(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility)

fn connection_moved(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility, invoker: Invoker)

fn connection_timeout(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId)

fn channel_announced(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId)

fn channel_description_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId)

fn channel_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, old_channel: Option<Channel>)

fn channel_created(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, invoker: Invoker)

fn channel_deleted(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, invoker: Invoker)

fn channel_edited(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, old_channel: Option<Channel>, invoker: Invoker)

fn channel_password_updated(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId)

fn channel_moved(&mut self, api: &mut TsApi, server_id: ServerId, channel_id: ChannelId, new_parent_channel_id: ChannelId, invoker: Invoker)

The current parent id of the channel is the old one, the new parent id is given as a parameter.

fn message(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker, target: MessageReceiver, message: String, ignored: bool) -> bool

A message was received. ignored describes, if the friend and fool system of TeamSpeak ignored the message. Return false if the TeamSpeak client should handle the message normally or true if the client should ignore the message.

fn poke(&mut self, api: &mut TsApi, server_id: ServerId, invoker: Invoker, message: String, ignored: bool) -> bool

A user poked us. ignored describes, if the friend and fool system of TeamSpeak ignored the message. Return false if the TeamSpeak client should handle the poke normally or true if the client should ignore the poke.

fn channel_kick(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, old_channel_id: ChannelId, new_channel_id: ChannelId, visibility: Visibility, invoker: Invoker, message: String)

fn server_kick(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, invoker: Invoker, message: String)

fn server_ban(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, invoker: Invoker, message: String, time: u64)

fn talking_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, talking: TalkStatus, whispering: bool)

The old values of talking and whispering are available from the connection. They will be updated after this functions returned.

fn avatar_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, path: Option<String>)

fn connection_channel_group_changed(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, channel_group_id: ChannelGroupId, channel_id: ChannelId, invoker: Invoker)

fn connection_server_group_added(&mut self, api: &mut TsApi, server_id: ServerId, connection: Invoker, server_group_id: ServerGroupId, invoker: Invoker)

fn connection_server_group_removed(&mut self, api: &mut TsApi, server_id: ServerId, connection: Invoker, server_group_id: ServerGroupId, invoker: Invoker)

fn playback_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, samples: &mut [i16], channels: i32)

The voice data is available as 16 bit with 48 KHz. The channels are packed (interleaved). The callbacks with audio data are called from another thread than the other functions.

fn post_process_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, connection_id: ConnectionId, samples: &mut [i16], channels: i32, channel_speaker_array: &[Speaker], channel_fill_mask: &mut u32)

The callbacks with audio data are called from another thread than the other functions.

fn mixed_playback_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, samples: &mut [i16], channels: i32, channel_speaker_array: &[Speaker], channel_fill_mask: &mut u32)

The callbacks with audio data are called from another thread than the other functions.

fn captured_voice_data(&mut self, api: &mut TsApi, server_id: ServerId, samples: &mut [i16], channels: i32, send: &mut bool) -> bool

The recorded sound from the current capture device. send is set if the audio data will be send to the server. This attribute can be changed in this callback. The return value of this function describes if the sound data was altered. Return true if the sound was changed and false otherwise. The callbacks with audio data are called from another thread than the other functions.

fn permission_error(&mut self, api: &mut TsApi, server_id: ServerId, permission_id: PermissionId, error: Error, message: String, return_code: String) -> bool

Return false if the TeamSpeak client should handle the error normally or true if the client should ignore the error.

fn shutdown(&mut self, api: &mut TsApi)

Called if the plugin is disabled (either by the user or if TeamSpeak is exiting).