Trait ts3plugin::Plugin

pub trait Plugin {
    fn init(&mut self) -> InitResult;
    fn shutdown(&mut self);

    fn request_autoload(&self) -> bool { ... }
    fn on_connect_status_change(&mut self, server: Server, new_status: ConnectStatus, error_number: Error) { ... }
    fn on_client_move(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, move_message: String) { ... }
    fn on_client_move_moved(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, move_message: String) { ... }
    fn on_client_move_timeout(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, move_message: String) { ... }
    fn on_client_move_subscription(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility) { ... }
    fn on_talk_status_change(&mut self, client: Connection, status: TalkStatus, is_received_whisper: bool) { ... }
    fn on_update_channel_edited(&mut self, channel: Channel, invoker: Connection) { ... }
    fn on_update_client(&mut self, client: Connection, invoker: Connection) { ... }
    fn on_text_message(&mut self, mode: TextMessageTargetMode, sender: Connection, receiver: Connection, message: String, ff_ignored: bool) -> bool { ... }
    fn on_current_server_connection_changed(&mut self, server: Server) { ... }
    fn on_new_channel(&mut self, channel: Channel, parent_channel: Channel) { ... }
    fn on_new_channel_created(&mut self, channel: Channel, parent_channel: Channel, invoker: Connection) { ... }
    fn on_delete_channel(&mut self, channel: Channel, invoker: Connection) { ... }
    fn on_channel_move(&mut self, channel: Channel, new_parent_channel: Channel, invoker: Connection) { ... }
    fn on_client_kick_from_channel(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, kick_message: String) { ... }
    fn on_client_kick_from_server(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, kick_message: String) { ... }
    fn on_server_group_client_added(&mut self, client: Connection, group: ServerGroup, invoker: Connection) { ... }
    fn on_server_group_client_deleted(&mut self, client: Connection, group: ServerGroup, invoker: Connection) { ... }
    fn configure(&mut self) { ... }
    fn register_plugin_id(&mut self, id: &str) { ... }
    fn command_keyword(&self) -> Option<String> { ... }
    fn process_command(&mut self, connection_id: u64, command: &str) -> bool { ... }
    fn info_title(&self) -> String { ... }
    fn info_data(&mut self, connection_id: u64, id: u64, item_type: ItemType) -> Option<String> { ... }
    fn init_menus(&mut self) -> Option<(Vec<MenuItem>, String)> { ... }
    fn init_hotkeys(&mut self) -> Option<Vec<Hotkey>> { ... }
    fn on_update_channel(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_client_ids(&mut self, connection_id: u64, unique_client_id: &str, client_id: u16, client_name: &str) { ... }
    fn on_client_ids_finished(&mut self, connection_id: u64) { ... }
    fn on_server_edited(&mut self, connection_id: u64, editor_id: u16, editor_name: &str, editor_unique_id: &str) { ... }
    fn on_server_updated(&mut self, connection_id: u64) { ... }
    fn on_server_error(&mut self, connection_id: u64, error_message: &str, error: Error, return_code: &str, extra_message: &str) -> bool { ... }
    fn on_server_stop(&mut self, connection_id: u64, shutdown_message: &str) { ... }
    fn on_connection_info(&mut self, connection_id: u64, client_id: u16) { ... }
    fn on_server_connection_info(&mut self, connection_id: u64) { ... }
    fn on_channel_subscribe(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_channel_subscribe_finished(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_channel_unsubscribe(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_channel_unsubscribe_finished(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_channel_description_update(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_channel_password_changed(&mut self, connection_id: u64, channel_id: u64) { ... }
    fn on_playback_shutdown_complete(&mut self, connection_id: u64) { ... }
    fn on_sound_device_list_changed(&mut self, mode_id: &str, play_or_cap: bool) { ... }
    fn on_edit_playback_voice_data(&mut self, connection_id: u64, client_id: u16, samples: Vec<i16>, channels: i32) { ... }
    fn on_user_logging_message(&mut self, log_message: &str, log_level: LogLevel, log_channel: &str, log_id: u64, log_time: &str, complete_log_string: &str) { ... }
    fn on_client_ban_from_server(&mut self, connection_id: u64, client_id: u16, old_channel_id: u64, new_channel_id: u64, visibility: Visibility, kicker_id: u16, kicker_name: &str, kicker_unique_id: &str, time: u64, kick_message: &str) { ... }
    fn on_client_poke_event(&mut self, connection_id: u64, from_client_id: u16, poker_name: &str, poker_unique_id: &str, message: &str, ff_ignored: bool) -> bool { ... }
    fn on_client_self_variable_update(&mut self, connection_id: u64, flag: i32, old_value: &str, new_value: &str) { ... }
    fn on_file_list(&mut self, connection_id: u64, path: &str, name: &str, size: u64, datetime: u64, typename: FileListType, incompletesize: u64, return_code: &str) { ... }
    fn on_file_list_finished(&mut self, connection_id: u64, channel_id: u64, path: &str) { ... }
    fn on_file_info(&mut self, connection_id: u64, channel_id: u64, name: &str, size: u64, datetime: u64) { ... }
    fn on_server_group_list(&mut self, connection_id: u64, server_group_id: u64, name: &str, typename: i32, icon_id: i32, save_db: bool) { ... }
    fn on_server_group_list_finished(&mut self, connection_id: u64) { ... }
    fn on_server_group_by_cliend_id(&mut self, connection_id: u64, name: &str, server_group_list: u64, client_database_id: u64) { ... }
    fn on_server_group_perm_list(&mut self, connection_id: u64, server_group_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_server_group_perm_list_finished(&mut self, connection_id: u64, server_group_id: u64) { ... }
    fn on_server_group_client_list(&mut self, connection_id: u64, server_group_id: u64, client_database_id: u64, client_name_id: &str, client_unique_id: &str) { ... }
    fn on_channel_group_list(&mut self, connection_id: u64, channel_group_id: u64, name: &str, typename: i32, icon_id: i32, save_db: bool) { ... }
    fn on_channel_group_list_finished(&mut self, connection_id: u64) { ... }
    fn on_channel_group_perm_list(&mut self, connection_id: u64, channel_group_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_channel_group_perm_list_finished(&mut self, connection_id: u64) { ... }
    fn on_channel_perm_list(&mut self, connection_id: u64, channel_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_channel_perm_list_finished(&mut self, connection_id: u64, channel_id: i32) { ... }
    fn on_client_perm_list(&mut self, connection_id: u64, client_database_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_client_perm_list_finished(&mut self, connection_id: u64, client_database_id: u64) { ... }
    fn on_channel_client_perm_list(&mut self, connection_id: u64, channel_id: u64, client_database_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_channel_client_perm_list_finished(&mut self, connection_id: u64, channel_id: u64, client_database_id: u64) { ... }
    fn on_server_permission_error(&mut self, connection_id: u64, error_message: &str, error: Error, return_code: &str, failed_permission_id: u32) -> bool { ... }
    fn on_permission_list_group_end_id(&mut self, connection_id: u64, group_end_id: u32) { ... }
    fn on_permission_list(&mut self, connection_id: u64, permission_id: u32, permission_name: &str, permission_description: &str) { ... }
    fn on_permission_list_finished(&mut self, connection_id: u64) { ... }
    fn on_permission_overview(&mut self, connection_id: u64, client_database_id: u64, overview_type: i32, overview_id1: u64, overview_id2: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32) { ... }
    fn on_permission_overview_finished(&mut self, connection_id: u64) { ... }
}

The trait that has to be implemented by a plugin. To enhance a library to a working TeamSpeak plugin you have to call the macro create_plugin! afterwards.

Examples

A fully working example that creates a plugin that does nothing:

#[macro_use]
extern crate ts3plugin;

use ts3plugin::*;

struct MyTsPlugin;

impl MyTsPlugin
{
    fn new() -> MyTsPlugin
    {
        MyTsPlugin
    }
}

impl Plugin for MyTsPlugin
{
    fn init(&mut self) -> InitResult
    {
        println!("Inited");
        InitResult::Success
    }

    fn shutdown(&mut self)
    {
        println!("Shutdown");
    }
}

create_plugin!("My Ts Plugin\0", "0.1.0\0", "My Name\0",
    "A wonderful tiny example plugin\0", ConfigureOffer::No, MyTsPlugin);

Required Methods

fn init(&mut self) -> InitResult

Custom code called right after loading the plugin.

fn shutdown(&mut self)

Custom code called right before the plugin is unloaded.

Note:

If your plugin implements a settings dialog, it must be closed and deleted here, else the TeamSpeak client will most likely crash (library removed but dialog from the library code is still used).

Provided Methods

fn request_autoload(&self) -> bool

Plugin requests to be always automatically loaded by the TeamSpeak client unless the user manually disabled it in the plugin dialog. The default value is false which means the plugin won't be loaded automatically.

fn on_connect_status_change(&mut self, server: Server, new_status: ConnectStatus, error_number: Error)

fn on_client_move(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, move_message: String)

fn on_client_move_moved(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, move_message: String)

fn on_client_move_timeout(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, move_message: String)

fn on_client_move_subscription(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility)

fn on_talk_status_change(&mut self, client: Connection, status: TalkStatus, is_received_whisper: bool)

fn on_update_channel_edited(&mut self, channel: Channel, invoker: Connection)

fn on_update_client(&mut self, client: Connection, invoker: Connection)

fn on_text_message(&mut self, mode: TextMessageTargetMode, sender: Connection, receiver: Connection, message: String, ff_ignored: bool) -> bool

ff_ignored is true if the friend/foe manager ignored that message.

If true is returned, the TeamSpeak client will ignore this message, if false is returned, it will be handled normally.

fn on_current_server_connection_changed(&mut self, server: Server)

Client changed current server connection handler.

fn on_new_channel(&mut self, channel: Channel, parent_channel: Channel)

After a connection has been established, all current channels on the

fn on_new_channel_created(&mut self, channel: Channel, parent_channel: Channel, invoker: Connection)

fn on_delete_channel(&mut self, channel: Channel, invoker: Connection)

fn on_channel_move(&mut self, channel: Channel, new_parent_channel: Channel, invoker: Connection)

fn on_client_kick_from_channel(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, kick_message: String)

fn on_client_kick_from_server(&mut self, client: Connection, old_channel: Channel, new_channel: Channel, visibility: Visibility, invoker: Connection, kick_message: String)

fn on_server_group_client_added(&mut self, client: Connection, group: ServerGroup, invoker: Connection)

fn on_server_group_client_deleted(&mut self, client: Connection, group: ServerGroup, invoker: Connection)

fn configure(&mut self)

Plugin might offer a configuration window. If the ConfigureOffer is ConfigureOffer::No, this function does not need to be implemented.

fn register_plugin_id(&mut self, id: &str)

If the plugin wants to use error return codes, plugin commands, hotkeys or menu items, it needs to register a command ID. This function will be automatically called after the plugin was initialized. This function is optional. If you don't use these features, this function can be omitted. Note the passed pluginID parameter is no longer valid after calling this function, so you must copy it and store it in the plugin.

fn command_keyword(&self) -> Option<String>

Plugin command keyword. Return None if not used.

fn process_command(&mut self, connection_id: u64, command: &str) -> bool

Plugin processes console command. Return true if plugin handled the command, false if not handled.

fn info_title(&self) -> String

Implement the following three functions when the plugin should display a line in the server/channel/client info.

fn info_data(&mut self, connection_id: u64, id: u64, item_type: ItemType) -> Option<String>

Dynamic content shown in the right column in the info frame. Check the parameter type if you want to implement this feature only for specific item types. Return None to have the client ignore the info data.

fn init_menus(&mut self) -> Option<(Vec<MenuItem>, String)>

Initialize plugin menus. This function is called after init and register_plugin_id. A pluginID is required for plugin menus to work. register_plugin_id must be implemented to use menus. If plugin menus are not used by a plugin, return None.

Returns

A list of MenuItems and the identifier for the menu icon.

fn init_hotkeys(&mut self) -> Option<Vec<Hotkey>>

Initialize plugin hotkeys. Hotkeys require register_plugin_id to be implemented. This function is automatically called by the client after init.

fn on_update_channel(&mut self, connection_id: u64, channel_id: u64)

fn on_client_ids(&mut self, connection_id: u64, unique_client_id: &str, client_id: u16, client_name: &str)

fn on_client_ids_finished(&mut self, connection_id: u64)

fn on_server_edited(&mut self, connection_id: u64, editor_id: u16, editor_name: &str, editor_unique_id: &str)

fn on_server_updated(&mut self, connection_id: u64)

fn on_server_error(&mut self, connection_id: u64, error_message: &str, error: Error, return_code: &str, extra_message: &str) -> bool

Return true if the client should ignore that error because it was handled by this plugin.

fn on_server_stop(&mut self, connection_id: u64, shutdown_message: &str)

fn on_connection_info(&mut self, connection_id: u64, client_id: u16)

fn on_server_connection_info(&mut self, connection_id: u64)

fn on_channel_subscribe(&mut self, connection_id: u64, channel_id: u64)

fn on_channel_subscribe_finished(&mut self, connection_id: u64, channel_id: u64)

fn on_channel_unsubscribe(&mut self, connection_id: u64, channel_id: u64)

fn on_channel_unsubscribe_finished(&mut self, connection_id: u64, channel_id: u64)

fn on_channel_description_update(&mut self, connection_id: u64, channel_id: u64)

fn on_channel_password_changed(&mut self, connection_id: u64, channel_id: u64)

fn on_playback_shutdown_complete(&mut self, connection_id: u64)

fn on_sound_device_list_changed(&mut self, mode_id: &str, play_or_cap: bool)

fn on_edit_playback_voice_data(&mut self, connection_id: u64, client_id: u16, samples: Vec<i16>, channels: i32)

fn on_user_logging_message(&mut self, log_message: &str, log_level: LogLevel, log_channel: &str, log_id: u64, log_time: &str, complete_log_string: &str)

fn on_client_ban_from_server(&mut self, connection_id: u64, client_id: u16, old_channel_id: u64, new_channel_id: u64, visibility: Visibility, kicker_id: u16, kicker_name: &str, kicker_unique_id: &str, time: u64, kick_message: &str)

fn on_client_poke_event(&mut self, connection_id: u64, from_client_id: u16, poker_name: &str, poker_unique_id: &str, message: &str, ff_ignored: bool) -> bool

Return true if the plugin handled this event and the client should ignore it.

fn on_client_self_variable_update(&mut self, connection_id: u64, flag: i32, old_value: &str, new_value: &str)

fn on_file_list(&mut self, connection_id: u64, path: &str, name: &str, size: u64, datetime: u64, typename: FileListType, incompletesize: u64, return_code: &str)

fn on_file_list_finished(&mut self, connection_id: u64, channel_id: u64, path: &str)

fn on_file_info(&mut self, connection_id: u64, channel_id: u64, name: &str, size: u64, datetime: u64)

fn on_server_group_list(&mut self, connection_id: u64, server_group_id: u64, name: &str, typename: i32, icon_id: i32, save_db: bool)

fn on_server_group_list_finished(&mut self, connection_id: u64)

fn on_server_group_by_cliend_id(&mut self, connection_id: u64, name: &str, server_group_list: u64, client_database_id: u64)

fn on_server_group_perm_list(&mut self, connection_id: u64, server_group_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_server_group_perm_list_finished(&mut self, connection_id: u64, server_group_id: u64)

fn on_server_group_client_list(&mut self, connection_id: u64, server_group_id: u64, client_database_id: u64, client_name_id: &str, client_unique_id: &str)

fn on_channel_group_list(&mut self, connection_id: u64, channel_group_id: u64, name: &str, typename: i32, icon_id: i32, save_db: bool)

fn on_channel_group_list_finished(&mut self, connection_id: u64)

fn on_channel_group_perm_list(&mut self, connection_id: u64, channel_group_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_channel_group_perm_list_finished(&mut self, connection_id: u64)

fn on_channel_perm_list(&mut self, connection_id: u64, channel_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_channel_perm_list_finished(&mut self, connection_id: u64, channel_id: i32)

fn on_client_perm_list(&mut self, connection_id: u64, client_database_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_client_perm_list_finished(&mut self, connection_id: u64, client_database_id: u64)

fn on_channel_client_perm_list(&mut self, connection_id: u64, channel_id: u64, client_database_id: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_channel_client_perm_list_finished(&mut self, connection_id: u64, channel_id: u64, client_database_id: u64)

fn on_server_permission_error(&mut self, connection_id: u64, error_message: &str, error: Error, return_code: &str, failed_permission_id: u32) -> bool

Return true if the plugin handled this event and the client should ignore it.

fn on_permission_list_group_end_id(&mut self, connection_id: u64, group_end_id: u32)

fn on_permission_list(&mut self, connection_id: u64, permission_id: u32, permission_name: &str, permission_description: &str)

fn on_permission_list_finished(&mut self, connection_id: u64)

fn on_permission_overview(&mut self, connection_id: u64, client_database_id: u64, overview_type: i32, overview_id1: u64, overview_id2: u64, permission_id: u32, permission_value: i32, permission_negated: i32, permission_skip: i32)

fn on_permission_overview_finished(&mut self, connection_id: u64)