Module ts3plugin::ts3interface

Statics

ts3functions
The api functions provided by TeamSpeak

Functions

ts3plugin_apiVersion

ts3plugin_author
The author of the plugin. Can be called before init.

ts3plugin_currentServerConnectionChanged

ts3plugin_description
The desription of the plugin. Can be called before init.

ts3plugin_init

ts3plugin_name
Unique name identifying this plugin. The result of this function has to be a null-terminated static string. Can be called before init.

ts3plugin_offersConfigure
If the plugin offers the possibility to be configured by the user.

ts3plugin_onChannelMoveEvent

ts3plugin_onClientKickFromChannelEvent

ts3plugin_onClientKickFromServerEvent

ts3plugin_onClientMoveEvent

ts3plugin_onClientMoveMovedEvent

ts3plugin_onClientMoveSubscriptionEvent

ts3plugin_onClientMoveTimeoutEvent

ts3plugin_onConnectStatusChangeEvent

ts3plugin_onDelChannelEvent

ts3plugin_onNewChannelCreatedEvent

ts3plugin_onNewChannelEvent

ts3plugin_onServerGroupClientAddedEvent

ts3plugin_onServerGroupClientDeletedEvent

ts3plugin_onTalkStatusChangeEvent

ts3plugin_onTextMessageEvent

ts3plugin_onUpdateChannelEditedEvent

ts3plugin_onUpdateClientEvent

ts3plugin_requestAutoload

ts3plugin_setFunctionPointers

ts3plugin_shutdown

ts3plugin_version
The version of the plugin. Can be called before init.