Crate ts3plugin

Reexports

pub use ts3plugin_sys::clientlib_publicdefinitions::*;

pub use ts3plugin_sys::plugin_definitions::*;

pub use ts3plugin_sys::public_definitions::*;

Modules

ts3interface

Macros

create_plugin!
Create a plugin. This macro has to be called once per library to create the function interface that is used by TeamSpeak.

Structs

Channel

Connection

PluginData
A struct that is used for the internal representation of the plugin

Server

ServerGroup

Ts3Functions
Functions exported to plugin from main binary

TsApi

Enums

Error

InitResult

Traits

Plugin
The trait that has to be implemented by a plugin. To enhance a library to a working TeamSpeak plugin you have to call the macro create_plugin! afterwards.