Trait ts3plugin::Plugin

pub trait Plugin {
    fn init(&mut self) -> InitResult;
    fn shutdown(&mut self);

    fn offers_configure(&self) -> ConfigureOffer { ... }
}

The trait that has to be implemented by a plugin. To enhance a library to a working TeamSpeak plugin you have to call the macro create_plugin! afterwards.

Examples

A fully working example that creates a plugin that does nothing:

#![feature(box_raw)]
#[macro_use]
extern crate ts3plugin;

use ts3plugin::*;

struct MyTsPlugin;

impl Plugin for MyTsPlugin
{
    fn init(&mut self) -> InitResult
    {
        println!("Inited");
        InitResult::Success
    }

    fn shutdown(&mut self)
    {
        println!("Shutdown");
    }
}

create_plugin!("My Ts Plugin\0", "0.1.0\0", "My Name\0",
    "A wonderful tiny example plugin\0", MyTsPlugin);

Required Methods

fn init(&mut self) -> InitResult

Custom code called right after loading the plugin.

fn shutdown(&mut self)

Custom code called right before the plugin is unloaded.

Note:

If your plugin implements a settings dialog, it must be closed and deleted here, else the TeamSpeak client will most likely crash (library removed but dialog from the library code is still used).

Provided Methods

fn offers_configure(&self) -> ConfigureOffer

Tell client if plugin offers a configuration window. If this function is not overwritten, ConfigureOffer::No is returned.